Refactor PageFrame
