+++ src/UI/Root/Components/PageFrame/index.ts
@@ -1,1 +1,1 @@
-export * from "./PageFrame";
+export { Provider as PageFrame } from "./Provider";
